+++ NEWS.md
@@ -4,6 +4,8 @@
 - Add featureData slot to Chromatograms class and add mz, precursorMz and 
   productMz methods for Chromatograms (issue
   [#289](https://github.com/lgatto/MSnbase/issues/289)) <2017-12-18 Mon>.
+- Add readSRMData function to read chromatographic data from SRM (MRM)
+  experiments; issue [#286](https://github.com/lgatto/MSnbase/issues/286) <2018-01-10 Wed>.
 
 ## Changes in version 2.5.3
 - MSnSet to/from SummarizedExperiment coercion (contributed by Arne