+++ NEWS.md
@@ -11,6 +11,8 @@
   <2018-01-11 Thu>
 
 ## Changes in version 2.5.3
+- Add filterPrecursorScan for `MSnExp` and `OnDiskMSnExp`;
+  closes issue #282 and PR #287 <2017-12-16 Sat>.
 - MSnSet to/from SummarizedExperiment coercion (contributed by Arne
   Smits in PR [#284](https://github.com/lgatto/MSnbase/issues/284)) <2017-12-17 Sun>
 - Fix inverted M/Z axis in plot3D,MSmap (reported by Sylvain